Update README
Builds out the README by moving around some of the existing content
into the appropriate sections as well as stubbing out some of the
sections that still need to be developed. Those sections are Creating a
survey, Filling out a survey, and Survey administration. Also, a brief
overview of the OMK server installation is added.